Extract MIXER command into its own file
