+++ src/dos/meson.build
@@ -34,6 +34,7 @@ libdos_sources = files(
     'program_loadrom.cpp',
     'program_ls.cpp',
     'program_mem.cpp',
+    'program_mixer.cpp',
     'program_more.cpp',
     'program_more_output.cpp',
     'program_mount.cpp',
