+++ vs/dosbox.vcxproj.filters
@@ -661,6 +661,9 @@
     <ClCompile Include="..\src\dos\program_mem.cpp">
       <Filter>src\dos</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\dos\program_mixer.cpp">
+      <Filter>src\dos</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\dos\program_more.cpp">
       <Filter>src\dos</Filter>
     </ClCompile>
